# web/public/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Nexus-Mail API
  version: 0.5.0
  description: |
    Nexus-Mail 当前对外接口文档（Phase 5 初始版）。
    覆盖身份认证、项目/订单、供应商资源、钱包与结算相关主链路。
servers:
  - url: /api/v1
    description: 当前前后端统一反向代理入口
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
...
    AdminDashboardSummary:
      type: object
      properties:
        users:
          type: object
          properties:
            total:
              type: integer
        orders:
          type: object
          properties:
            total:
              type: integer
            waiting_email:
              type: integer
            ready:
              type: integer
            finished:
              type: integer
            canceled:
              type: integer
            timeout:
              type: integer
            completion_rate_bps:
              type: integer
              description: 完成订单数 / 总订单数，单位 bps（10000 = 100%）。
            timeout_rate_bps:
              type: integer
              description: 超时订单数 / 总订单数，单位 bps。
            cancel_rate_bps:
              type: integer
              description: 取消订单数 / 总订单数，单位 bps。
            gross_revenue:
              type: integer
              description: 已完成订单 final_price 合计，单位分。
            average_finished_order_value:
              type: integer
              description: 已完成订单平均 final_price，单位分。
        disputes:
          type: object
          properties:
            total:
              type: integer
            open:
              type: integer
            resolved:
              type: integer
            rejected:
              type: integer
            dispute_rate_bps:
              type: integer
              description: 争议单数量 / 总订单数，单位 bps。
        projects:
          type: object
          properties:
            total:
              type: integer
            active:
              type: integer
            inactive:
              type: integer
        suppliers:
          type: object
          properties:
            total:
              type: integer
        audit:
          type: object
          properties:
            total:
              type: integer
            create:
              type: integer
            revoke:
              type: integer
            success:
              type: integer
            denied_invalid:
              type: integer
            denied_scope:
              type: integer
            denied_whitelist:
              type: integer
            denied_rate_limit:
              type: integer
            denied_total:
              type: integer
            denied_rate_bps:
              type: integer
              description: 最近审计样本中的鉴权拒绝事件占比，单位 bps。
        supplier_settlements:
          type: object
          properties:
            pending_amount:
              type: integer
    AdminAuditEntry:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        api_key_id:
          type: integer
        action:
          type: string
          enum:
            - create
            - update_whitelist
            - update_risk_rule
            - admin_wallet_adjustment
            - update_supplier_cost_profile
            - resolve_dispute
            - revoke
            - success
            - denied_invalid
            - denied_scope
            - denied_whitelist
            - denied_rate_limit
        actor_type:
          type: string
        note:
          type: string
        created_at:
          type: string
          format: date-time
    AdminOverviewResponse:
      type: object
      properties:
        generated_at:
          type: string
          format: date-time
        summary:
          $ref: '#/components/schemas/AdminDashboardSummary'
        recent_audit:
          type: array
          items:
            $ref: '#/components/schemas/AdminAuditEntry'
...
security:
  - bearerAuth: []
paths:
...
  /admin/overview:
    get:
      summary: 获取管理员运营仪表盘深度统计
      description: 返回管理员 overview 聚合数据，包括订单率、流水、争议率、API Key 鉴权拒绝率以及最近审计事件。
      responses:
        '200':
          description: 管理员仪表盘统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminOverviewResponse'
        '401':
          description: 缺少或无效的用户认证令牌
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: 非管理员无权访问
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /admin/audit:
    get:
      summary: 管理员审计日志查询
      description: 支持按 action、actor_type、user_id、api_key_id 与 limit 过滤；action 覆盖 API Key 生命周期、鉴权拒绝、风控规则更新以及 Phase 4 财务运营动作（admin_wallet_adjustment、update_supplier_cost_profile、settle_supplier_pending、resolve_dispute）。
      parameters:
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: actor_type
          schema:
            type: string
            enum: [user, system, admin]
        - in: query
          name: user_id
          schema:
            type: integer
        - in: query
          name: api_key_id
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
            maximum: 200
      responses:
        '200':
          description: 审计日志列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/AdminAuditEntry'
        '400':
          description: 过滤条件不受支持
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...